requirements.txt
Django==2.2.3
celery==4.3.0
djangorestframework==3.10.0
markdown==3.1.1
django-filter==2.2.0
django-allauth==0.39.1
django-rest-auth==0.9.5
django-celery-beat==1.5.0
django-celery-results==1.1.2
asgiref==3.1.4
